Update README.md
Based on a case study of our choice, we identified and produced a project management plan for a technical/software project that, as a team, we selected. The project was in the initiation stage and was implementing change and adding new services or products.We were able to establish specific project objectives. The project contained requirements for different project resources, including people. It was important to balance the selection of a project with sufficient complexity and size to clearly demonstrate the team’s understanding of and ability to apply the project management concepts, models, tools, and techniques that we have studied in this course. The project chosen was currently underway and should not be a completed project.

Tasks:
1) Develop an Agile project charter
2) Define the Scope of the project; developing requirements, a scope statement, a work breakdown structure, and a scope baseline.
3) Identify appropriate people for Agile roles
4) Project Scheduling; Define Increments and Timeboxes to keep a project on track
5) Manage a Solution Development Team
6) Communicate with business stakeholders in a project environment
7) Use Agile techniques, including prioritization of requirements, to ensure on-time project delivery,
8) Present the benefits of Agile approaches to senior management
9) Identify risks to delivery of projects managed using the Agile approach